third_party/blink/web_tests/platform/generic/http/tests/loading/preload-image-sizes-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded('resources/image2.png'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?1'); is false
PASS internals.isPreloaded('resources/base-image1.png?1'); is false
PASS internals.isPreloaded('resources/base-image2.png?1'); is false
PASS internals.isPreloaded('resources/base-image3.png?1'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?2'); is false
PASS internals.isPreloaded('resources/base-image1.png?2'); is false
PASS internals.isPreloaded('resources/base-image2.png?2'); is false
PASS internals.isPreloaded('resources/base-image3.png?2'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?3'); is false
PASS internals.isPreloaded('resources/base-image1.png?3'); is false
PASS internals.isPreloaded('resources/base-image2.png?3'); is false
PASS internals.isPreloaded('resources/base-image3.png?3'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?4'); is false
PASS internals.isPreloaded('resources/base-image1.png?4'); is false
PASS internals.isPreloaded('resources/base-image2.png?4'); is false
PASS internals.isPreloaded('resources/base-image3.png?4'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?5'); is false
PASS internals.isPreloaded('resources/base-image1.png?5'); is false
PASS internals.isPreloaded('resources/base-image2.png?5'); is false
PASS internals.isPreloaded('resources/base-image3.png?5'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?6'); is false
PASS internals.isPreloaded('resources/base-image1.png?6'); is false
PASS internals.isPreloaded('resources/base-image2.png?6'); is false
PASS internals.isPreloaded('resources/base-image3.png?6'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?7'); is false
PASS internals.isPreloaded('resources/base-image1.png?7'); is false
PASS internals.isPreloaded('resources/base-image2.png?7'); is false
PASS internals.isPreloaded('resources/base-image3.png?7'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?8'); is false
PASS internals.isPreloaded('resources/base-image1.png?8'); is false
PASS internals.isPreloaded('resources/base-image2.png?8'); is false
PASS internals.isPreloaded('resources/base-image3.png?8'); is true
PASS internals.isPreloaded('resources/preload-test.jpg?9'); is false
PASS internals.isPreloaded('resources/base-image1.png?9'); is false
PASS internals.isPreloaded('resources/base-image2.png?9'); is false
PASS internals.isPreloaded('resources/base-image3.png?9'); is true
PASS internals.isPreloaded('resources/dup-image1.png?7'); is false
PASS internals.isPreloaded('resources/dup-image1.png?8'); is false
PASS internals.isPreloaded('resources/dup-image2.png?8'); is false
PASS internals.isPreloaded('resources/dup-image3.png?8'); is false
PASS internals.isPreloaded('resources/preload-test.jpg?10'); is true
PASS internals.isPreloaded('resources/base-image1.png?10'); is false
PASS internals.isPreloaded('resources/base-image2.png?10'); is false
PASS internals.isPreloaded('resources/base-image3.png?10'); is false
PASS successfullyParsed is true

TEST COMPLETE

          
